+++ docs/changelog.rst
@@ -4,8 +4,8 @@ Changelog
 3.0.0-beta.5 (unreleased)
 -------------------------
 
+- bugfix: `skip_row()` fix crash when model has m2m field and none is provided in upload (#1439)
 - Fix deprecation in example application: Added support for transitional form renderer (#1451)
-
 
 3.0.0-beta.4 (2022-05-17)
 -------------------------